requirements.txt
transformers==4.29.2
datasets==2.12.0
torch==2.0.1
accelerate==0.19.0
evaluate==0.4.0
safetensors==0.3.1
Pillow==9.5.0
timm==0.9.2
scipy==1.10.1
requests==2.25.1
ipykernel==6.23.1
scikit-learn==1.2.2 
matplotlib==3.7.1
natten==0.14.6
# pip install trectools